requirements.txt
# Core Flask and web framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23

# Authentication
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
bcrypt==4.0.1

# CORS support
Flask-CORS==4.0.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# OpenAI integration
openai==1.3.5

# Data processing
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code quality
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==21.2.0

# Development tools
Flask-Migrate==4.0.5
